+++ ATM/bank.h
@@ -12,12 +12,12 @@ struct Record {
     string ano;
     vector<string> cnos;
     vector<string> pins;
-    long double balance{ 0.0 };
+    long int balance{ 0 };
 };
 
 class File {
 public:
-    void balup(string ano, int mon);
+    void balup(string ano, long int mon);
     vector<Record>& contain();
     void write(vector<Record> Data);
     string getacc(string cno);
@@ -31,7 +31,7 @@ public:
     void operations();
     void insert();
     int balance();
-    void withdraw(int mon);
-    void deposit(int mon);
-    void transfer(string ano, int mon); 
+    void withdraw(long int mon);
+    void deposit(long int mon);
+    void transfer(string ano, long int mon); 
 };
